Moved tests
Interface tests are unit tests
